requirements.txt
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.6.1
eventlet==0.36.0
fastapi==0.110.0
flower==2.0.1
fuzzywuzzy==0.18.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
humanize==4.9.0
idna==3.6
iniconfig==2.0.0
kombu==5.3.5
Levenshtein==0.25.0
lxml==5.1.0
Mako==1.3.2
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pip-review==1.3.0
pluggy==1.4.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.0
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.7.0
redis==5.0.3
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.16
starlette==0.36.3
tabulate==0.9.0
tornado==6.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
